refactor: moved stats logic to separate component
