+++ src/components/history/HistoryContainer.js
@@ -4,7 +4,7 @@ import { UserContext } from '../providers/UserProvider';
 import HistoryList from './HistoryList';
 import { Input, Typography, Button, Grid, Card, useTheme } from '@material-ui/core';
 import Dropdowns from '../inputform/Dropdowns';
-import { Doughnut } from 'react-chartjs-2';
+import StatsContainer from '../stats/StatsContainer';
 
 export default function HistoryContainer(props) {
 	const { entries, userEntries, getUserEntries, deleteEntry } = React.useContext(EntryContext);
@@ -12,30 +12,11 @@ export default function HistoryContainer(props) {
 	const [ searchTerm, setSearchTerm ] = useState('');
 
 	const theme = useTheme();
-	const { palette: { type, primary, secondary, error } } = theme;
+	const { palette: { type/* , primary, secondary, error */ } } = theme;
 
 	const inputStyle = {
 		color: type === 'light' ? 'primary' : 'secondary'
 	};
-
-	const initialLegendOptions = {
-		display: true,
-		position: 'bottom',
-		fullWidth: true,
-		reverse: false,
-		labels: {
-			// fontColor: type === 'light' ? primary.main : secondary.main,
-			fontFamily: 'Roboto',
-			fontSize: 16,
-			fontStyle: 'bold'
-			// usePointStyle: true,
-		}
-	};
-	const totalClosed = useRef(0);
-	const totalOpen = useRef(0);
-	const totalSession = useRef(null);
-	const totalBreak = useRef(0);
-	const legendOptions = useRef(initialLegendOptions);
 
 	const initialFilter = {
 		isCompleted: false,
@@ -129,74 +110,6 @@ export default function HistoryContainer(props) {
 		[ activeUser.id ]
 	);
 
-	useEffect(
-		() => {
-			totalOpen.current = userEntries.reduce(function(acc, object) {
-				if (!object.isCompleted) {
-					return acc + 1;
-				}
-				return acc;
-			}, 0);
-			totalClosed.current = userEntries.length - totalOpen.current;
-			totalSession.current = userEntries.reduce(function(acc, object) {
-				return acc + object.totalWorkTime;
-			}, 0);
-			totalBreak.current = userEntries.reduce(function(acc, object) {
-				return acc + object.totalBreakTime;
-			}, 0);
-		},
-		[ userEntries ]
-	);
-
-	const data = {
-		labels: [ 'OPEN', 'CLOSED' ],
-		datasets: [
-			{
-				data: [ totalOpen.current, totalClosed.current ],
-				backgroundColor: [ error.main, secondary.main ],
-				hoverBackgroundColor: [ error.dark, secondary.dark ],
-				borderColor: 'none',
-				borderWidth: 0
-			}
-		]
-	};
-
-	const workDoughnutOptions = {
-		maintainAspectRatio: true
-		// title: {
-		//     display: true,
-		//     fontFamily: 'Roboto',
-		//     fontSize: 16,
-		//     text: ["SESSION v. BREAK TOTALS", "(minutes)"],
-		// }
-	};
-
-	const data2 = {
-		labels: [ 'SESSION', 'BREAK' ],
-		datasets: [
-			{
-				data: [ Math.floor(totalSession.current / 60000), Math.floor(totalBreak.current / 60000) ],
-				backgroundColor: [ primary.light, secondary.light ],
-				hoverBackgroundColor: [ primary.main, secondary.main ],
-				borderColor: 'none',
-				borderWidth: 0
-			}
-		]
-	};
-
-	const sessionDisplayHours = Math.floor((totalSession.current % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60));
-	const sessionDisplayMinutes = Math.floor((totalSession.current % (1000 * 60 * 60)) / (1000 * 60));
-
-	const breakDisplayHours = Math.floor((totalBreak.current % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60));
-	const breakDisplayMinutes = Math.floor((totalBreak.current % (1000 * 60 * 60)) / (1000 * 60));
-
-	// const totalTime = totalSession.current + totalBreak.current
-	const totalDisplayHours = sessionDisplayHours + breakDisplayHours;
-	const totalDisplayMinutes = sessionDisplayMinutes + breakDisplayMinutes;
-
-	const sessionBreakRatio =
-		totalSession.current && totalSession.current / (totalSession.current + totalBreak.current);
-
 	return (
 		<Grid container spacing={1}>
 			<Grid item xs={6} sm={3}>
@@ -260,77 +173,7 @@ export default function HistoryContainer(props) {
 				<Typography variant="h6" style={{ textAlign: 'center' }}>
 					STATS
 				</Typography>
-				<Card style={{ textAlign: 'center', margin: '0 10px', padding: '10px' }}>
-					<div style={{ textAlign: 'center' }}>
-						<Typography>Keep up the great work! #DevEveryDay</Typography>
-						<Card style={{ padding: '10px', margin: '10px' }}>
-							<Typography style={{ fontSize: '1.5em' }}>BUG TOTALS</Typography>
-							<Doughnut
-								data={data}
-								legend={legendOptions.current}
-								width={200}
-								height={200}
-								options={{ maintainAspectRatio: true }}
-							/>
-						</Card>
-						<Typography>Open: {totalOpen.current}</Typography>
-						<Typography>Closed: {totalClosed.current}</Typography>
-						<Typography>Total: {userEntries.length}</Typography>
-						<hr width="80%" />
-						{sessionBreakRatio > 0.8333 ? (
-							<React.Fragment>
-								<Typography>Take frequent debugging breaks to stay sharp.</Typography>
-								<Typography>
-									<span
-										style={{
-											color: type === 'light' ? error.main : error.light,
-											fontWeight: '900',
-											fontSize: '1.4em'
-										}}
-									>
-										{100 - Math.floor(sessionBreakRatio * 100)}%
-									</span>{' '}
-									of your time has been dedicated to breaks.
-								</Typography>
-							</React.Fragment>
-						) : (
-							<React.Fragment>
-								<Typography>You have mastered the Art of the Break. #DevSourcerer</Typography>
-								<Typography>
-									<span
-										style={{
-											color: type === 'light' ? primary.main : secondary.main,
-											fontWeight: '900',
-											fontSize: '1.4em'
-										}}
-									>
-										{100 - Math.floor(sessionBreakRatio * 100)}%
-									</span>{' '}
-									of your time has been dedicated to breaks.
-								</Typography>
-							</React.Fragment>
-						)}
-						<Card style={{ padding: '10px', margin: '10px' }}>
-							<Typography style={{ fontSize: '1.5em' }}>SESSION v. BREAK TOTALS (minutes)</Typography>
-							<Doughnut
-								data={data2}
-								legend={legendOptions.current}
-								width={200}
-								height={200}
-								options={workDoughnutOptions}
-							/>
-						</Card>
-						<Typography>
-							Session Time: {sessionDisplayHours}h {sessionDisplayMinutes}m
-						</Typography>
-						<Typography>
-							Break Time: {breakDisplayHours}h {breakDisplayMinutes}m
-						</Typography>
-						<Typography>
-							Total: {totalDisplayHours}h {totalDisplayMinutes}m
-						</Typography>
-					</div>
-				</Card>
+                <StatsContainer />		
 			</Grid>
 			{/* </React.Fragment> */}
 		</Grid>
